dreams/requirements.txt
# (venv:tutorial)~/tutorial$ pip install -U -r requirements.txt
Django
django-registration-redux
ipython
psycopg2
selenium

